Implement aggegates for Simple RGC Transduction
