+++ src/main/kotlin/simplergc/services/batch/output/BatchCsvColocalizationOutput.kt
@@ -1,40 +1,11 @@
 package simplergc.services.batch.output
 
-import kotlin.math.pow
-import kotlin.math.sqrt
+import simplergc.services.Aggregate
 import simplergc.services.AggregateRow
 import simplergc.services.CsvTableWriter
-import simplergc.services.DoubleField
-import simplergc.services.Field
-import simplergc.services.IntField
 import simplergc.services.Parameters
 import simplergc.services.Table
 import simplergc.services.colocalizer.output.CsvColocalizationOutput
-
-class CsvAggregateGenerator(val values: List<Int>) : AggregateGenerator() {
-
-    private fun computeStandardDeviation(): Double {
-        val squareOfMean = values.average().pow(2)
-        val meanOfSquares = values.map { it.toDouble().pow(2) }.average()
-        return sqrt(meanOfSquares - squareOfMean)
-    }
-
-    override fun generateMean(): Field<*> {
-        return DoubleField(values.average())
-    }
-
-    override fun generateStandardDeviation(): Field<*> {
-        return DoubleField(computeStandardDeviation())
-    }
-
-    override fun generateStandardErrorOfMean(): Field<*> {
-        return DoubleField(computeStandardDeviation() / sqrt(values.size.toDouble()))
-    }
-
-    override fun generateCount(): Field<*> {
-        return IntField(values.size)
-    }
-}
 
 /**
  * Outputs multiple CSVs into an output folder.
@@ -56,11 +27,10 @@ class BatchCsvColocalizationOutput(transductionParameters: Parameters.Transducti
 
     override fun generateAggregateRow(
         aggregate: Aggregate,
-        fileValues: List<List<Int>>
+        rawValues: List<List<Int>>,
+        spaces: Int
     ): AggregateRow {
-        return AggregateRow(aggregate.abbreviation, fileValues.map { values ->
-            aggregate.generateValue(CsvAggregateGenerator(values))
-        })
+        return colocalizationOutput.generateAggregateRow(aggregate, rawValues, spaces)
     }
 
     override fun output() {
