+++ src/main/kotlin/simplergc/services/batch/output/BatchXlsxColocalizationOutput.kt
@@ -1,37 +1,13 @@
 package simplergc.services.batch.output
 
 import org.apache.poi.xssf.usermodel.XSSFWorkbook
+import simplergc.services.Aggregate
 import simplergc.services.AggregateRow
-import simplergc.services.Field
-import simplergc.services.FormulaField
 import simplergc.services.Parameters
 import simplergc.services.Table
 import simplergc.services.TableWriter
 import simplergc.services.XlsxTableWriter
 import simplergc.services.colocalizer.output.XlsxColocalizationOutput
-
-class XlsxAggregateGenerator(column: Char, numCells: Int) : AggregateGenerator() {
-
-    private val startCellRow = 2
-    private val endCellRow = numCells + startCellRow - 1
-    private val cellRange = "$column$startCellRow:$column$endCellRow"
-
-    override fun generateMean(): Field<*> {
-        return FormulaField("AVERAGE($cellRange)")
-    }
-
-    override fun generateStandardDeviation(): Field<*> {
-        return FormulaField("STDEV($cellRange)")
-    }
-
-    override fun generateStandardErrorOfMean(): Field<*> {
-        return FormulaField("STDEV($cellRange)/SQRT(COUNT($cellRange))")
-    }
-
-    override fun generateCount(): Field<*> {
-        return FormulaField("COUNT($cellRange)")
-    }
-}
 
 /**
  * Outputs single XLSX file with multiple sheets.
@@ -54,16 +30,10 @@ class BatchXlsxColocalizationOutput(transductionParameters: Parameters.Transduct
 
     override fun generateAggregateRow(
         aggregate: Aggregate,
-        fileValues: List<List<Int>>
+        rawValues: List<List<Int>>,
+        spaces: Int
     ): AggregateRow {
-        var column = 'B'
-        val rowValues = mutableListOf<Field<*>>()
-        fileValues.forEach { values ->
-            rowValues.add(aggregate.generateValue(XlsxAggregateGenerator(
-                column++, values.size
-            )))
-        }
-        return AggregateRow(aggregate.abbreviation, rowValues)
+        return colocalizationOutput.generateAggregateRow(aggregate, rawValues, spaces)
     }
 
     override fun output() {
